changed the data type for the price to bigdecimal

+++ api/src/main/java/api/ProductDto.java
@@ -2,14 +2,16 @@ package api;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductDto {
     private Long id;
     private String title;
-    private int price;
+    private BigDecimal price;
     private String categoryTitle;
 
-    public ProductDto(Long id, String title, int price, String categoryTitle) {
+    public ProductDto(Long id, String title, BigDecimal price, String categoryTitle) {
         this.id = id;
         this.title = title;
         this.price = price;
